Introduce testing
As a testing strategy we'll generate apps that pretend to be buildpacks using the "inline" buildpack. We can then use this along with Hatchet to test our buildpack behavior.

Adding in these tests exposed a problem with the previous logic. There was a bug where if a Procfile existed but didn't contain a `release:` designation then the runtime check wouldn't be executed.