[close #1] Add leaky build path detection
The build dir and runtime dir are not the same on Heroku so a path like `/tmp/build_asdflasldkfjalsdkj` might be used at build time but at runtime it would be `/app`. It's easy to accidentally put a build path in the PATH at built time and not realize it. This feature detects and errors on this possible failure mode.